FC_RAG/requirements.txt
# requirements.txt
python-dotenv
pdfplumber
openai
chromadb
pytesseract
Pillow
PyPDF2
